Feat: 페이지 컴포넌트 코드 정리

+++ week4/week4-assignment/src/pages/Modal/DeleteIdModal.tsx
@@ -23,6 +23,10 @@ const DeleteIdModal = ({ onClose }: DeleteIdModalProps) => {
   }
 
   const handleDeleteId = async () => {
+    if (!userId) {
+      alert("사용자 정보가 유효하지 않습니다.");
+      return;
+    }
     try {
       const response = await deleteId(userId);
       alert("회원 삭제 성공");
@@ -35,15 +39,15 @@ const DeleteIdModal = ({ onClose }: DeleteIdModalProps) => {
 
   return ReactDOM.createPortal(
     <div className={modalBackgroundStyle}>
-      <div className={modalContentStyle}>
+      <div
+        className={modalContentStyle}
+        role="dialog"
+        aria-modal="true"
+        aria-labelledby="modal-title"
+      >
         <h2>정말 탈퇴하시겠어요?</h2>
         <p>탈퇴 후에는 모든 정보가 삭제돼요</p>
-        <div
-          className={modalBtnContainer}
-          role="dialog"
-          aria-modal="true"
-          aria-labelledby="modal-title"
-        >
+        <div className={modalBtnContainer}>
           <button onClick={onClose}>취소</button>
           <button onClick={handleDeleteId}>회원 탈퇴</button>
         </div>
